rhbz1215274 - enable min doc percent for file type project
